### 2024-09-25 #7282 @editorialbot
[PRE REVIEW]: respdetect: A Matlab tool for detecting breath events from whale biologger data
**Submitting author:** @ashleyblawas (<a href="http://orcid.org/0000-0003-4109-9003">Ashley Blawas</a>)
**Repository:** https://github.com/ashleyblawas/respdetect
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
