+++ paper/paper.bib
@@ -5,7 +5,7 @@
 	volume = {25},
 	number = {2},
 	pages = {327-350},
-	doi = {https://doi.org/10.1111/j.1748-7692.2008.00255.x},
+	doi = {10.1111/j.1748-7692.2008.00255.x},
 	year = {2009}
 }
 
@@ -17,7 +17,8 @@
   number={10},
   pages={jeb243872},
   year={2022},
-  publisher={The Company of Biologists Ltd}
+  publisher={The Company of Biologists Ltd},
+  doi = {10.1242/jeb.243872}
 }
 
 
