### 2024-09-27 #7282 @kthyng
@ashleyblawas
* any plans to make a matlab package for more seamless installation?
* maybe make it more obvious there is a wiki with a link in your readme
* The "plotting" in the wiki didn't have a plot?
* test directory just seems to have .mat files — no test code?
