+++ README.md
@@ -20,7 +20,7 @@ MATLAB toolbox dependencies:
 4. Statistics and Machine Learning Toolbox
 
 ### Installation Instructions
-You may either directly download this repository and add it to your MATLAB path or clone it using: git clone https://github.com/ashleyblawas/respdetect.git.
+You may either 1) directly download this repository or 2) clone it using: git clone https://github.com/ashleyblawas/respdetect.git.
 
 When you open the `main.m` file you will see that the first step steps up your directories. This setup is detailed in the repository Wiki [here](https://github.com/ashleyblawas/respdetect/wiki/3.-Setup). You will need to edit a `paths.txt` file which you will be prompted to select to designate the path of your tools and your data. After running the first step of `main/.m,` the `respdetect` tools will be on your path and no further installation is needed.
 
