### 2024-10-01 5aaa3df
Update paths.txt
