### 2024-10-23 #7282 @ashleyblawas
- This toolbox can be installed directly by downloading the repository and adding its path to Matlab so it does not save the user significant effort to instead install through a Matlab package
- I added a more obvious link to the Wiki in the Usage section of the readme
- This can been updated to add the plots that were missing!
- These test files are intended to be manually run through the series of processing steps. There is a built-in function within main.m to automatically compare the outputs of the two manual tests to the validation files also included in the test folder
