### 2024-10-23 #7282 @kthyng
> These test files are intended to be manually run through the series of processing steps. There is a built-in function within main.m to automatically compare the outputs of the two manual tests to the validation files also included in the test folder

Is this meant to be described in step 7 in the wiki? If so, are the subsequent steps 7a and 7b missing, which would walk the user through the testing to see if the code is working correctly?
