### 2024-10-23 #7282 @kthyng
@ashleyblawas
* can you also check those DOIs listed above
* also there are some inline reference issues in your paper (check example paper for help https://joss.readthedocs.io/en/latest/example_paper.html#example-paper)
