+++ paper/paper.md
@@ -40,6 +40,6 @@ These tools are were intially developed to detect breaths from DTAGs deployed on
 
 # Acknowledgements
 
-I would like to acknowledge Drs. Doug Nowacek and Andy Read for their support of this tool development, the Duke Superpod for project feedback and recommendations, Dr. Jeanne Shearer for pre-processing the short-finned pilot whale record used for testing, and Dr. Dave Haas, PMC-BS, and Socioambiental Aonsultores Ltd. for sharing and pre-processing the humpbaack whales record used for testing. This work was supported by the Duke University Marine Laboratory.
+I would like to acknowledge Drs. Doug Nowacek and Andy Read for their support of this tool development, the Duke Superpod for project feedback and recommendations, Dr. Jeanne Shearer for pre-processing the short-finned pilot whale record used for testing, and Dr. Dave Haas, PMC-BS, and Socioambiental Consultores Ltd. for sharing and pre-processing the humpbaack whales record used for testing. This work was supported by the Duke University Marine Laboratory.
 
 # References
