+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 # Statement of Need
 Physiological investigations of free-swimming cetaceans have, historically, been limited because of these animals’ large body size and fully aquatic lifestyle [@Block:2005; @Ponganis:2007; @Hooker:2007]. Recent developments in digital bio-logging devices have enabled direct measurements of some physiological parameters in free-swimming cetaceans [@Czapanskiy:2022; @Goldbogen:2019; @Williams:2017], but such investigations remain challenging due to constraints on the design and deployment of these devices. Our ability to measure physiological parameters in free-swimming cetaceans is critical to understanding how these species will respond to disturbance and environmental change. Therefore, development of novel methods to extract physiological information from bio-logging sensors can advance the fields of vertebrate physiology and conservation biology. 
 
-Respiratory parameters, including respiratory rate, are of particular interest in understanding the diving physiology of cetaceans because of their role in determining gas exchange and, therefore, the oxygen supply that supports breath-hold diving. Previous studies have determined respiratory rate of free-swimming cetaceans by direct observation [@Sumich:1983; @Blix:1995; @Williams:2009; @Christansen;2014], from acoustic signals in tags equipped with hydrophones [@Goldbogen:2008], and from patterns in the pitch and depth signals in accelerometry tags [@Miller:2010; @Roos:2016;, @Goldbogen:2008]). These methods are viable in some species, particularly those for which a breath is taken during each surfacing event, but they are not applicable for others that exhibit long periods of time at the surface (i.e., logging) with multiple breathing events, such as short-finned pilot whales. 
+Respiratory parameters, including respiratory rate, are of particular interest in understanding the diving physiology of cetaceans because of their role in determining gas exchange and, therefore, the oxygen supply that supports breath-hold diving. Previous studies have determined respiratory rate of free-swimming cetaceans by direct observation [@Sumich:1983; @Blix:1995; @Williams:2009; @Christansen:2014], from acoustic signals in tags equipped with hydrophones [@Goldbogen:2008], and from patterns in the pitch and depth signals in accelerometry tags [@Miller:2010; @Roos:2016;, @Goldbogen:2008]). These methods are viable in some species, particularly those for which a breath is taken during each surfacing event, but they are not applicable for others that exhibit long periods of time at the surface (i.e., logging) with multiple breathing events, such as short-finned pilot whales. 
 
 # Summary
 Here, I present `respdetect`, a set of Matlab tools to detect breathing events from high-resolution movement data recorded by digital acoustic recording tags (DTAGs). This novel method takes advantage of the high flow rates and large tidal volumes that are typical of cetacean breaths and detects the movement artifacts resulting from a breathing event. The method employs several signal processing steps followed by peak detection based on [@ Czapanskiy:2022]’s method for ballistocardiography from accelerometer data collected from blue whales to derive movement parameters from acceleration and pitch data and identify movement artifacts associated with respirations. The library of functions associated with resp_detect allow users to identify all breaths in the tag record and export the timing of each breath.
