### 2024-11-25 #7282 @ashleyblawas
Not sure if this is regarding the initial DOI issues on Sept. 25th but those are all corrected now. The inline reference issues have also been corrected.
