+++ paper/paper.bib
@@ -84,9 +84,8 @@
 	doi = {10.1242/jeb.023366},
 	language = {en},
 	number = {23},
-	urldate = {2019-07-15},
 	journal = {Journal of Experimental Biology},
-	author = {Goldbogen, Jeremy A. and Calambokidis, John and Croll, Donald A. and Harvey, James T. and Newton, Kelly M. and Oleson, Erin M. and Schorr, Greg and Shadwick, Robert E.},
+	author = {Goldbogen, J. A. and Calambokidis, John and Croll, Donald A. and Harvey, James T. and Newton, Kelly M. and Oleson, Erin M. and Schorr, Greg and Shadwick, Robert E.},
 	month = dec,
 	year = {2008},
 	pmid = {19011211},
