### 2024-12-02 #7282 @kthyng
Please check that your inline vs. parenthetical references are correct. You can compare with the example paper here: https://joss.readthedocs.io/en/latest/example_paper.html — line 32.

> This has been addressed to make it more clear how the user should use Step 7 for testing, including a more thorough description of Steps 7a & 7b.

I see there is some description. Jumping there and reading it, I wouldn't know how to do the testing, but I'll let the reviewers comment on it if they choose to do so.
