### 2025-01-02 #7282 @mikemahoney218
Happy to take this one on!

I should be able to start pinging people for reviews early next week.
