### 2025-04-09 ashleyblawas/respdetect#1 @smnnlt
Adopt a function based workflow
I would suggest to restructure the main.m into separate functions that can be run in isolation. For example, create `load_data()`, `make_metadata()`, `find_dives()`, `process_data()`, `detect_breaths()`, `plot_breaths()` and another function, maybe `run_all()` combining all of these. For users of your software, it is much easier to work with functions with documented input and output then to work with one big analysis script.

This would make it easier to run the analysis sequentially, but also allows to rerun parts of the analysis (e.g., with different parameters). It also eliminates the need to clear variables after each code section, but you would need a new way of handling variables that are currently kept constant across the script (e.g., the paths).

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
