### 2025-04-09 ashleyblawas/respdetect#2 @smnnlt
Recreate path selection
I do not really understand the current process of setting the data and tools paths, where you have to write the paths in a `.txt` file of a particular structure which is then imported into Matlab to extract the pathes as variables. In my view, it would be much easier to require the user to set the paths as a variable at the beginning of the script (or even better as function arguments).

Maybe it is also possible to auto-detect the tools paths with `which`. This would make make it easier to get the software started as only the data path has to be given. However, you should then implement the option to manually select the tool paths as a fall back if the auto-detection fails.

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
