### 2025-04-09 ashleyblawas/respdetect#3 @smnnlt
Use input validation
Particularly important if you rely more on functions, but also relevant to the current status of the project is input validation. Following programming best practices, you should check if any user input is what you expect it to be.

Matlab has its own system of input validation (https://mathworks.com/help/matlab/matlab_prog/function-argument-validation-1.html). You could complement this with simple `if` checks and informative error messages, e.g., to check that the `dive_thres` argument is not just a number but also non-negative.

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
