### 2025-04-09 ashleyblawas/respdetect#4 @smnnlt
Create function documentations
At the moment, the relevant function arguments are well explained at the beginning of the script but have to be set somewhere down in the respective sections. I strongly suggest that when converting main.m to functions (https://github.com/ashleyblawas/respdetect/issues/1) you systematically document each function and its parameters.

Matlab has a particular syntax for documenting functions that has the additional benefit that you can run `help` to get the function details: https://mathworks.com/help/matlab/matlab_prog/add-help-for-your-program.html

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
