### 2025-04-09 ashleyblawas/respdetect#5 @smnnlt
Follow Matlab syntax best practices
Matlab displays several syntax warnings in main.m and most other scripts. While they do not change the code's results, you should strive to eliminate these warnings when you produce software code for others.

The exact type of warning is very diverse, but most warnings relate to a lack of preallocation in loops or unused variables.

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
