### 2025-04-09 ashleyblawas/respdetect#6 @smnnlt
More informative testing system
I could not reproduce the analysis results as the manual test you described failed. This could be because I might have used different parameters for dive and breath detection than the ones used for generating the test results. You should make it very clear which input parameters you used for creating your testing results.

It would also be good to provide more informative messages for test failures. If the number of breaths is different from the test data, this line currently fails: https://github.com/ashleyblawas/respdetect/blob/298703de2af6d0125cefb72449a77a61ba213fce/respdetect/main.m#L963C1-L963C44
This should lead to a more informative failure message that the number of breaths is not identical between the data and the test data.

You should also at least consider using automated testing for (parts of) your software: https://www.mathworks.com/help/matlab/matlab-unit-test-framework.html

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
