### 2025-04-09 ashleyblawas/respdetect#7 @smnnlt
Broken links in wiki
The wiki is well-written and explains the basic usage of the package to new users in a accessible way. However, on the wiki main page, the links to section S1 and S2 seem to be broken.

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
