### 2025-04-09 ashleyblawas/respdetect#8 @smnnlt
Export to R missing?
Maybe I am missing something, but where can I find the export to R referenced in Wiki section S1?

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
