### 2025-04-09 ashleyblawas/respdetect#9 @smnnlt
Add contributing details
As per JOSS submission criteria, your software should provide contribution details:

> [Community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)
>
> There should be clear guidelines for third-parties wishing to:
>
>  - Contribute to the software
>
>  - Report issues or problems with the software
>
>  - Seek support

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
