### 2025-04-09 ashleyblawas/respdetect#10 @smnnlt
Include dtagtools code
Since `dtagtools` is released under a GPL license, you could consider including its source code in your software. This would ease getting started, since only one repo would have to be downloaded instead of two. It would also make it easier to auto-detect or manually provide the script paths (https://github.com/ashleyblawas/respdetect/issues/2).

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
