### 2025-04-09 ashleyblawas/respdetect#11 @smnnlt
Minor modification to the JOSS paper
I really like the software paper and the way it demonstrates the software's benefits to a wide audience. I have some minor comments:

- The paper is currently inconsistent in naming the software (`respdetect` vs. `resp_detect`)
- 37: I would include a citation to dtagtools here
- 47: Remove "are"

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
