### 2025-04-09 #7858 @smnnlt
Hi @ashleyblawas, thank you very much for creating this software and giving me the opportunity to review it.

I have finished my first round of review, see the issues opened in the repo and my review checklist. I tested the software on Windows 11 with Matlab 2024b installed.

As a short disclaimer, I have no experience in marine biology, so I am not qualified to make judgements about the correctness of the methods applied. But I have written research software and work in the field of human exercise physiology, so I feel that I am able to comment on some aspects.

I believe there are some major modifications that should be undertaken to improve this project in terms of usability and functionality. At the moment, the project feels more like a well-documented analysis script rather than research software. This is not a bad thing per se and totally fine for individual research projects. But if you want to make this work accessible to a wider user base there are several consideration you should make, see the issues I opened. And I definitely think it is worth going these extra miles, given the relevance of your software and the work you have already put into it.

Feel free to ask me any questions regarding my comments in this thread or the respective issues.
