### 2025-04-20 ashleyblawas/respdetect#12 @sfregosi
Suggested minor edits to the manuscript
To better meet the animal research guidelines:
- Suggest adding a short paragraph stating the availability of two test datasets and providing some basic background (location, date, species, tag type) on the two tags, or at least a citation to where data from these tags may be published elsewhere
- Either in the paragraph describing the test data or in the acknowledgements, provide relevant permitting and IACUC information for the tagging work
- Also include this information in the REAMDME.txt that is in each test tag folder

Line edits/comments
- L35 - 'resp_detect' should be respdetect in same font as line 28?
- L41 - same as L35
- L49 - same as L35
- Figure 1 - Suggest adding more detail to the figure caption better describing each panel

openjournals/joss-reviews#7858
