### 2025-04-20 ashleyblawas/respdetect#13 @sfregosi
Bug - 'nanmedian' conflict if not using MATLAB 2020a
### **Problem:**
I used MATLAB 2021a for my testing. Because of this I got a warning on line 589 of `main.m`:

`Warning: Name is nonexistent or not a directory: C:\Program Files\MATLAB\R2020a\toolbox\stats\stats `

I first ignored this warning but then got an error on line 602:

```
Error using nanmedian
Too many input arguments.
Error in mad (line 49)
    y = nanmedian(abs(x - nanmedian(x,dim)),dim);
Error in mvksdensity>apply_censoring_get_bandwidth (line 334)
            sig = mad(ty,1,1) / 0.6745;
Error in mvksdensity (line 105)
[ty,~,weight,u,foldpoint,maxp] = apply_censoring_get_bandwidth(cens,yData,ty,n,ymax,weight,u,d);
Error in ksdensity (line 226)
    [fout,xout,u,plottype] = mvksdensity(yData,xi,varargin{:});
Error in detect_peaks (line 15)
        [f_d,xi_d] = ksdensity(dist);
Error in main (line 602)
    [j_locs, j_width, j_prom, idx, rm_group] = detect_peaks(metadata.fs, jerk_smooth, 3,
    min_sec_apart);
```

I was able to trace this back to that warning. Both MATLAB's stats toolbox and the `dtagtools` have a function `nanmedian`. Because L589 did not run (which would have put the stats toolbox `nanmedian` higher on the path) the script was trying to use the `dtagtools` version of `nanmedian` which only takes one input argument.

```
>> which('-all', 'nanmedian')
C:\Users\selene\Documents\MATLAB\dtagtools\dtag2\nanmedian.m
C:\Program Files\MATLAB\R2021a\toolbox\stats\distributed\@distributed\nanmedian.m  % distributed method
C:\Program Files\MATLAB\R2021a\toolbox\stats\stats\nanmedian.m                     % Shadowed
```

### **Temporary fix:**
I manually changed the year of line 589 and was able to continue on through testing because that moved the correct `nanmedian` to the top of the path:

```
>> which('-all', 'nanmedian')
C:\Program Files\MATLAB\R2021a\toolbox\stats\stats\nanmedian.m
C:\Program Files\MATLAB\R2021a\toolbox\stats\distributed\@distributed\nanmedian.m  % distributed method
C:\Users\selene\Documents\MATLAB\dtagtools\dtag2\nanmedian.m                       % Shadowed
```

### **Permanent fix:**
- If `dtagtools` is packaged directly with `respdetect` as suggested in #10 then the `dtagtools` version of `nanmedian` could be removed/renamed
- Alternatively, could search for the correct path and erplace L589 with something like below, which should be more robust to different installation paths

```
    all_nanmedians = which('-all', 'nanmedian');	% cell array of all nanmedian.m locations
    positions = regexpi(all_nanmedians, 'toolbox[/\\]stats[/\\]stats'); % find "toolbox/stats/stats"
    index = find(~cellfun('isempty', positions), 1); % pick the one find() found
    statsDir = fileparts(all_nanmedians{index});	% extract the directory name
    addpath(statsDir) % add stats path to top

```

RE: openjournals/joss-reviews#7858
