### 2025-04-20 ashleyblawas/respdetect#14 @sfregosi
Add option to re-run steps 2-4 even if that .mat already exists
In testing I found I wanted to re-run the metadata, dive detection, and movement steps (Sections 2 to 4) with different parameters. If I just tried to re-run those sections, the script spits out a statement that the table already exists and to move to the next step.

I was able to force re-make the tables by deleting them. However, I think it would be more user-friendly to have an option to re-create any of those tables from the script (without having to delete), or, ideally within that function following implementation of #1.

Suggest modifying the if/else statements that starts at L137, L228, and L330 that check for an existing table to provide an option to recreate the table or not, similar to how it prompts for a `y/n` to initially make the table.

RE: openjournals/joss-reviews#7858
