### 2025-04-20 ashleyblawas/respdetect#15 @sfregosi
Wiki edits
### 3. Setup

- It states that the user will be prompted to select tags but I was never prompted - it just defaulted to the single tag in the test folder

### 4. Creating data files

- Throughout this page you reference that the created files will end with `-` and then some letters but none of them actually have the dash. Minor, but I'd suggest removing all the dashes.
- It would be very helpful to provide the tag metadata (on/off times and tag type version) for testing to ensure the user sets up the metadata table properly (I tried saying guessing one was a CATS tag and got an error in `main.m` on L826). Either provide this info in this section, or better, provide a link to the wiki page 7 and provide the details there. This info should also be in the manuscript (see #12).
- L300 (Step 3- dives) triggers the MATLAB 'beep'. This may just be personal preference/an individual response, but because that sound usually means some sort of error/issue I started to go digging for where something when wrong. I don't think the beep is necessary to indicate the process is complete since a message is printed to the Command Window when it is done. But, if it is important to leave in then perhaps add some text to the Wiki stating that a beep will indicate when the process is done (so testers know this is normal behavior).
- 'Frequency' is misspelled in Step 4.

### 6. Plotting breath detections
- In the introduction paragraph, 'identifies surfacings as periods where the depth is > 5 m' could be confusing. Suggest changing to '...depth is shallower than the set dive minimum threshold (set in Step 3, default is 5 m)'. And the next sentence '...ascends from below that dive threshold and descends back below that threshold with < 10 seconds...'. Is the 10 second cut off hard coded or also something a user can change, similar to the dive threshold?

### 7. Test detector
- Broadly, I would suggest pulling this section out of `main.m` to be a standalone `validate.m` script. This script would assume the process of `main.m` was already run so would load those detected breaths, and then compare them to the provided validated breaths
- I was not able to complete testing this section because I got an error on line 963 because the test and validated arrays were not the same size (same error as #6). The wiki states that this means 'something has gone wrong' but the wiki doesn't provide any guidance to troubleshoot the problem.

RE: openjournals/joss-reviews#7858
