### 2025-04-29 ashleyblawas/respdetect#10 @sfregosi
I agree with this issue. By using this approach, it would also provide a more straightforward way to solve #13.
