### 2025-04-29 ashleyblawas/respdetect#2 @sfregosi
I also had some trouble navigating the existing path set up process. I've added my comments below:

The current path set up with paths.txt is not super robust to mistakes

- The paths.txt I downloaded had `"##` to start the header lines (different than wiki screenshot, also different than standard MATLAB comment line). Either make wiki and paths.txt consistent, or better suggest changing to `%` to match MATLAB. Also add a line to the paths.txt that specifies how to mark a 'comment' line.
- Reading in the path values of paths.txt relies on exact spacing and number of header lines. I initially just commented out your default paths so I'd have a reference to them, so I had extra header lines and the tools, mat, and data paths were not read in correctly in `main.m` lines 76-78. Alternatively, you could use `eval` (although use of `eval` is controversial, I think this is a place in which it might make sense) and parse in these paths directly in the paths.txt file. I think this would be more flexible to differing spacing and number of lines.

My suggested improvements above would not be necessary by implementing #1. I wholly agree with @smnnlt 's suggestion to have these path as function input arguments.
