### 2025-04-29 ashleyblawas/respdetect#3 @sfregosi
I second this issue and believe it is necessary to really make this useable by others.
