### 2025-04-29 ashleyblawas/respdetect#4 @sfregosi
I agree this is important when implementing #1.

Some additional resources:

- There is a function on FileExchange that provides a template for the top of any new functions: https://www.mathworks.com/matlabcentral/fileexchange/27132-automatic-template-for-new-functions
- I have my own (less elegant) version that may help as a guide: https://github.com/sfregosi/myUtils/blob/main/newFun.m
