### 2025-04-29 #7858 @sfregosi
@ashleyblawas,

I've completed my initial review of `respdetect`. Overall, I think this is a useful toolset that complements existing `dtagtools` nicely and is an important contribution to the cetacean biologging field. I appreciate your work to share this with the community in an open and reproducible way. While I was able to get it mostly working as it stands, I agree with @smnnlt that it would be greatly improved if it is 'functionized.' I also identified some small bugs and areas the documentation could be improved to better guide a new user. I have provided comments as either new issues on the `respdetect` repository or as comments on existing issues. I am happy to provide clarity on or discuss any of these issues and would be happy to test a revised version in the future.

Best,
Selene
