### 2025-04-30 ashleyblawas/respdetect#1 @smnnlt
> As an alternative to creating a `run_all()` function, I would suggest creating a `run_all` example script that would serve as the 'backbone' of running this type of analysis. This script could be the location where a user would define their relevant paths (removing the need for the paths text file) that could then be used as input arguments to each function. I would envision that as a user, I would create one of these scripts for each experiment/tag deployment which would keep a record of the parameters used, etc., but would not require modifying any of the original package code.

I support this suggestion, I think it is a really good idea to have such an example script.
