### 2025-06-30 #7858 @ashleyblawas
Yes, thank you. I will respond to these comments within the week.

Best,
Ashley

On Tue, Jun 24, 2025 at 1:53 PM Kristen Thyng ***@***.***>
wrote:

> *kthyng* left a comment (openjournals/joss-reviews#7858)
> <https://github.com/openjournals/joss-reviews/issues/7858#issuecomment-3001827841>
>
> @ashleyblawas <https://github.com/ashleyblawas> Can you please comment on
> the reviewer comments?
>
