### 2025-07-08 #7858 @ashleyblawas
It will be this week. Apologizes for the delay and thanks for the reminder!On Jul 8, 2025, at 11:24 AM, Kristen Thyng ***@***.***> wrote:﻿kthyng left a comment (openjournals/joss-reviews#7858)
@ashleyblawas Do you have an estimate for how long you will need to complete a response? If it will be longer than this week, I will pause this issue (add a label for bookkeeping). If it is going to take longer than a month, we can discuss that next.
