### 2025-07-10 ashleyblawas/respdetect#7 @ashleyblawas
The links have been updated and this issue is now fixed.
