### 2025-07-10 f1ce4e5
Create CONTRIBUTING.md
