+++ CONTRIBUTING.md
@@ -1,0 +1,45 @@
+## Welcome
+
+Pull requests, bug reports, and all other forms of contribution are welcomed and encouraged!
+
+## Code of Conduct
+
+Before contributing, please read the [Contributor Covenant Code of Conduct](https://www.contributor-covenant.org/version/2/1/code_of_conduct/) and be sure to adhere to these expectations when making a contribution to the repository. 
+
+## Before You Start
+
+Before you start contributing, ensure the following:
+
+* Please do not debug your specific project in GitHub issues (instead use it for filing bugs and feature requests). If you have a question specific to your application, please contact me via email at ashleyblawas@gmail.com.
+* Check that you are using the latest version of the project as not to work on outdated code.
+* Search through existing issues to see if your issue has previously been reported. If so, comment with any additional information you have and/or indicate that you have this problem too which will make it higher priority to fix.
+
+## Contribution Workflow
+
+### Fork and clone repository
+
+GitHub provides a useful guide to forking and cloning which you can find [here](https://docs.github.com/en/desktop/adding-and-cloning-repositories/cloning-and-forking-repositories-from-github-desktop).
+
+### Report issues and bugs
+
+ - Double check that this issue has not already been reported.
+   
+ - Provide sufficient detail to allow the bug to be reproduced including system details (OS, MATLAB version, etc.) as well as a full description and screenshots/videos of the bug when applicable.
+
+### Pull requests
+
+*Note: All contributions will be licensed under the project's license.*
+
+- Limit pull request to one per bug fix or feature. When doing so, do not reformat code that is unrelated to your change.
+
+- For larger changes, try to open an issue to discuss a strategy with the repository maintainer to ensure this doesn't duplicate ongoing efforts.
+
+- Include the ability to test new functionality when possible.
+
+- Keep coding styles and conventions as consistent as possible.
+
+- Add documentation. Document your changes with code doc comments or in existing guides.
+
+---
+
+Adapted from [The Good Docs Project](https://thegooddocsproject.dev/) and in part from the [CONTRIBUTING.md](https://github.com/jessesquires/.github/blob/main/CONTRIBUTING.md) written by @jessesquires.
