### 2025-07-10 ashleyblawas/respdetect#9 @ashleyblawas
A CONTRIBUTING.md file has been added that addresses the these points.
