### 2025-07-10 bd71326
Update paper.md
