+++ paper/paper.md
@@ -26,20 +26,20 @@ Physiological investigations of free-swimming cetaceans have, historically, been
 Respiratory parameters, including respiratory rate, are of particular interest in understanding the diving physiology of cetaceans because of their role in determining gas exchange and, therefore, the oxygen supply that supports breath-hold diving. Previous studies have determined respiratory rate of free-swimming cetaceans by direct observation [@Sumich:1983; @Blix:1995; @Williams:2009; @Christiansen:2014], from acoustic signals in tags equipped with hydrophones [@Goldbogen:2008], and from patterns in the pitch and depth signals in accelerometry tags [@Miller:2010; @Roos:2016; @Goldbogen:2008]. These methods are viable in some species, particularly those for which a breath is taken during each surfacing event, but they are not applicable for others that exhibit long periods of time at the surface (i.e., logging) with multiple breathing events, such as short-finned pilot whales. 
 
 # Summary
-Here, I present `respdetect`, a set of Matlab tools to detect breathing events from high-resolution movement data recorded by digital acoustic recording tags (DTAGs). This novel method takes advantage of the high flow rates and large tidal volumes that are typical of cetacean breaths and detects the movement artifacts resulting from a breathing event. The method employs several signal processing steps followed by peak detection based on @Czapanskiy:2022’s method for ballistocardiography from accelerometer data collected from blue whales to derive movement parameters from acceleration and pitch data and identify movement artifacts associated with respirations. The library of functions associated with resp_detect allow users to identify all breaths in the tag record and export the timing of each breath.
+Here, I present `respdetect`, a set of Matlab tools to detect breathing events from high-resolution movement data recorded by digital acoustic recording tags (DTAGs). This novel method takes advantage of the high flow rates and large tidal volumes that are typical of cetacean breaths and detects the movement artifacts resulting from a breathing event. The method employs several signal processing steps followed by peak detection based on @Czapanskiy:2022’s method for ballistocardiography from accelerometer data collected from blue whales to derive movement parameters from acceleration and pitch data and identify movement artifacts associated with respirations. The library of functions associated with respdetect allow users to identify all breaths in the tag record and export the timing of each breath.
 
-`Respdetect` is intended to be used with existing DTAG processing tools developed by Mark Johnson. These tools require the same folder structure necessary for DTAG data processing. A DTAG record should be processed by exporting the prh.m file which contains the accelerometer and magentometer data transformed into the whale frame, as well as the animal’s pitch, roll, and heading. To use resp_detect tools, the user is asked to specify the deployment prh files to analyze. After identifying dives and calculating movement parameters, the user then executes the breath detector which will identify all breathing events in the tag record.
+`Respdetect` is intended to be used with existing DTAG processing tools developed by Mark Johnson. These tools require the same folder structure necessary for DTAG data processing. A DTAG record should be processed by exporting the prh.m file which contains the accelerometer and magentometer data transformed into the whale frame, as well as the animal’s pitch, roll, and heading. To use respdetect tools, the user is asked to specify the deployment prh files to analyze. After identifying dives and calculating movement parameters, the user then executes the breath detector which will identify all breathing events in the tag record.
 
-![An example of respdetect breath detections from a short-finned pilot whale DTAG2 record.](../images/gm08_143b_allbreaths_ex.jpg)
[…]
