### 2025-07-10 2a1749e
Update paper.md
