### 2025-07-10 298f665
Update paper.md to add tag info
