+++ paper/paper.md
@@ -38,6 +38,8 @@ Following detection, the tools allow the user to visualize the respiratory rate 
 
 These tools are were intially developed to detect breaths from DTAGs deployed on short-finned pilot whales, but  should be useful for any DTAG or CATS tag record. The best use cases for respdetect are for tags that were deployed close to the blowhole. 
 
+There are two test datasets avaiable for users to validate nominal function of `respdetect.` The short-finned pilot whale record (*gm08_143b*) was collected in May of 2008 in Cape Hatteras, North Carolina, USA using a DTAG 2 (D2) and has been previously included in published studies including [@Shearer:2022]. The humpback whale record (*mn17_310a*) was collected in November of 2017 of the coast of Brazil using a D2.
+
 # Acknowledgements
 
 I would like to acknowledge Drs. Doug Nowacek and Andy Read for their support of this tool development, the Duke Superpod for project feedback and recommendations, Dr. Jeanne Shearer for pre-processing the short-finned pilot whale record used for testing, and Dr. Dave Haas, PMC-BS, and Socioambiental Consultores Ltd. for sharing and pre-processing the humpback whales record used for testing. This work was supported by the Duke University Marine Laboratory. Data was collected under NOAA Permits 14241 and 14809 and approved by the Institutional Animal Care and Use Committee of Duke University.
