### 2025-07-10 bbef860
Update paper.bib
