+++ paper/paper.bib
@@ -1,3 +1,19 @@
+
+@article{Shearer:2022,
+	title = {Short-finned pilot whales exhibit behavioral plasticity in foraging strategies mediated by their environment},
+	volume = {695},
+	issn = {0171-8630, 1616-1599},
+	url = {https://www.int-res.com/abstracts/meps/v695/p1-14/},
+	doi = {10.3354/meps14132},
+	language = {en},
+	journal = {Marine Ecology Progress Series},
+	author = {Shearer, Jeanne M. and Jensen, Frants H. and Quick, Nicola J. and Friedlaender, Ari and Southall, Brandon and Nowacek, Douglas P. and Bowers, Matthew and Foley, Heather J. and Swaim, Zachary T. and Waples, Danielle M. and Read, Andrew J.},
+	month = aug,
+	year = {2022},
+	pages = {1--14},
+}
+
+
 @article{Williams:2009,
 	author = {Williams, Rob and Noren, Dawn P.},
 	title = {Swimming speed, respiration rate, and estimated cost of transport in adult killer whales},
