### 2025-07-10 ashleyblawas/respdetect#12 @ashleyblawas
Thank you for this suggestion - these changes have been made.
