### 2025-07-10 538d7ac
Update paper.md
