### 2025-07-10 57237f3
Update paper.md
