+++ paper/paper.md
@@ -28,7 +28,7 @@ Respiratory parameters, including respiratory rate, are of particular interest i
 # Summary
 Here, I present `respdetect`, a set of Matlab tools to detect breathing events from high-resolution movement data recorded by digital acoustic recording tags (DTAGs). This novel method takes advantage of the high flow rates and large tidal volumes that are typical of cetacean breaths and detects the movement artifacts resulting from a breathing event. The method employs several signal processing steps followed by peak detection based on @Czapanskiy:2022’s method for ballistocardiography from accelerometer data collected from blue whales to derive movement parameters from acceleration and pitch data and identify movement artifacts associated with respirations. The library of functions associated with respdetect allow users to identify all breaths in the tag record and export the timing of each breath.
 
-`Respdetect` is intended to be used with existing DTAG processing tools developed by Mark Johnson. These tools require the same folder structure necessary for DTAG data processing. A DTAG record should be processed by exporting the prh.m file which contains the accelerometer and magentometer data transformed into the whale frame, as well as the animal’s pitch, roll, and heading. To use respdetect tools, the user is asked to specify the deployment prh files to analyze. After identifying dives and calculating movement parameters, the user then executes the breath detector which will identify all breathing events in the tag record.
+`Respdetect` is intended to be used with existing DTAG processing tools developed by Mark Johnson [@DTAGTools]. These tools require the same folder structure necessary for DTAG data processing. A DTAG record should be processed by exporting the prh.m file which contains the accelerometer and magentometer data transformed into the whale frame, as well as the animal’s pitch, roll, and heading. To use respdetect tools, the user is asked to specify the deployment prh files to analyze. After identifying dives and calculating movement parameters, the user then executes the breath detector which will identify all breathing events in the tag record.
 
 ![An example of respdetect breath detections from a short-finned pilot whale DTAG2 record. The first panel shows the time depth record (black line) for this segment of data with detected breaths plotted using squares. The next three panels show smoothed Shannon entropy derived from surge, jerk, and pitch, respectively, with corresponding squares at timepoints corresponding to each breath.](../images/gm08_143b_allbreaths_ex.jpg)
 
@@ -36,7 +36,7 @@ Following detection, the tools allow the user to visualize the respiratory rate 
 
 <img align="right" src="../images/gm08_143b_resprate_ex.jpg" alt="Continuous respiratory rate from breath detections by respdetect for a short-finned pilot whale before and after deep dives." width="500"/>
 
-These tools are were intially developed to detect breaths from DTAGs deployed on short-finned pilot whales, but  should be useful for any DTAG or CATS tag record. The best use cases for respdetect are for tags that were deployed close to the blowhole. 
+These tools were intially developed to detect breaths from DTAGs deployed on short-finned pilot whales, but  should be useful for any DTAG or CATS tag record. The best use cases for respdetect are for tags that were deployed close to the blowhole. 
 
 There are two test datasets avaiable for users to validate nominal function of `respdetect.` The short-finned pilot whale record (*gm08_143b*) was collected in May of 2008 in Cape Hatteras, North Carolina, USA using a DTAG 2 (D2) and has been previously included in published studies including [@Shearer:2022]. The humpback whale record (*mn17_310a*) was collected in November of 2017 of the coast of Brazil using a D2.
 
