### 2025-07-10 d8723c7
Update paper.bib
