+++ paper/paper.bib
@@ -1,4 +1,14 @@
 
+@misc{johnson_stacyderuiterdtagtools_2013,
+	title = {stacyderuiter/dtagtools},
+	copyright = {GPL-3.0},
+	url = {https://github.com/stacyderuiter/dtagtools},
+	author = {Johnson, Mark},
+	month = nov,
+	year = {2013},
+	note = {original-date: 2013-11-06T09:50:48Z},
+}
+
 @article{Shearer:2022,
 	title = {Short-finned pilot whales exhibit behavioral plasticity in foraging strategies mediated by their environment},
 	volume = {695},
