### 2025-07-10 ashleyblawas/respdetect#11 @ashleyblawas
These modifications have been made, thank you for the suggestions!
