### 2025-07-10 8a413e3
Update paper.bib
