+++ paper/paper.bib
@@ -1,5 +1,5 @@
 
-@misc{johnson_stacyderuiterdtagtools_2013,
+@misc{DTAGTools,
 	title = {stacyderuiter/dtagtools},
 	copyright = {GPL-3.0},
 	url = {https://github.com/stacyderuiter/dtagtools},
