### 2025-07-11 e26eccb
Functionalize the path setting and data loading
