### 2025-07-11 ashleyblawas/respdetect#2 @ashleyblawas
I have created functions to set the paths and load the data, so this should now be resolved. The original intent was to have the paths set outside the script in a file that would not be edited regularly as to minimize the user accidently changing the paths. I think building this into functions is much better, so appreciate the ideas!
