### 2025-07-11 d7fd00d
Removing paths.txt as now obsolete
