### 2025-07-11 c52a7d7
Changed how saving metdata and dropping save_metadata
