+++ respdetect/make_metadata.m
@@ -1,15 +1,96 @@
-%%  Function to create metadata structure so you don't have to deal with lots of variables
-
-function [metadata] = make_metadata(tag, recdir, prefix, fs, tag_ver, acousaud_filename, time_tagon, time_tagoff);
-
-% Create metadata variable
-metadata.tag = tag;
-metadata.prefix = prefix;
-metadata.recdir = recdir;
-metadata.fs = fs;
-metadata.tag_ver = tag_ver;
-metadata.tagon_time = time_tagon;
-metadata.tagoff_time = time_tagoff;
-metadata.acousticaudit_filename = acousaud_filename;
-
-
+function make_metadata(taglist, dataPath)
+    arguments
+        taglist (1, :) cell
+        dataPath (1,:) char
+    end
+    % Loops over taglist and creates metadata files if not present.
+    %
+    % Syntax:
+    %   process_tag_metadata(taglist, dataPath)
+    %
+    % Inputs:
+    %   taglist  - Cell array of tag names
+    %   dataPath - Base path to data (e.g., 'C:\my_data\')
+    
+    % Author: Ashley Blawas
+    % Last Updated: 7/11/2025
+    % Stanford University
+    
+    
+    for k = 1:length(taglist);
+        
+        % Save tag name
+        if exist('k','var') == 1
+            tag = taglist{k};
+        else
+            tag = taglist{1};
+        end
+        
+        speciesCode = tag(1:2);
+        
+        % Make metadata file
+        metadata_fname = strcat(dataPath, speciesCode, "\metadata\", tag, "md.mat");
+        
+        % If it already exists don't make it again
+        if isfile(metadata_fname) == 1
+            fprintf("A metadatafile exists for %s - go the next section!\n", tag)
+        else % Otherwise, make and save the metadata file
+            fprintf("No metadata file exists for %s.\n", tag)
+            str = input("Do you want to make a metadata file now? (y/n)\n",'s');
+            if strcmp(str, "y") == 1
+                
+                %Set path for prh files
+                settagpath('prh',strcat(dataPath, speciesCode,'\prh'));
+                
+                % Load the tag's prh file
+                loadprh(tag);
+                
+                %Print out the fs so you can check it's what you expect
+                fprintf('fs = %i Hz\n', fs);
+                
+                % Calculate time variables from full duration of tag deployment
+                [time_sec, time_min, time_hour] = calc_time(fs, p);
+                
+                % Print out duration of tag
+                totalSeconds = max(time_sec); % Convert to total seconds
+                
+                % Convert to hours, minutes, seconds
+                hours    = floor(totalSeconds / 3600);
+                minutes  = floor(mod(totalSeconds, 3600) / 60);
+                sec  = round(mod(totalSeconds, 60));
+                
+                % Display formatted record duration
+                fprintf('Record duration (hh:mm:ss): %.0f:%2.0f:%02d.\n', ...
+                    hours, minutes, sec);
+                
+                % Designate tag on and off time
+                [tag_on] = get_tag_on(time_sec, p, fs);
+                [tag_off] = get_tag_off(time_sec, p, fs);
+                
+                % Calculate tag on duration
+                tag_dur = datestr(seconds(tag_off-tag_on),'HH:MM:SS');
+                
+                % Display formatted tag on duration
+                fprintf('Tag on duration (hh:mm:ss): %s.\n', ...
+                    tag_dur);
+                
+                % Designate the tag version
+                prompt = {'Enter DTAG version (D2 or D3 or CATS):'};
+                dlgtitle = 'Input';
+                dims = [1 35];
+                definput = {'CATS'};
[…]
