### 2025-07-11 db9d091
Update to work with functions through metadata
