### 2025-07-11 86bf506
Delete save_metadata because now obsolete
