### 2025-07-11 ee745e2
Consistently format functions
