### 2025-07-11 67461ef
Moving the movement code into functions
