+++ respdetect/calc_move.m
@@ -1,131 +1,155 @@
-% Calculate movement metrics
-
-function calc_move(fs, Aw, p, data_path, tag, pitch, roll, head)
- %Calculate filtered acceleration
- 
-            % Rename Aw vector
-            surge = Aw(:, 1);
-            sway = Aw(:, 2);
-            heave = Aw(:, 3);
-            
-            % Make temp variables without NaNs to work with and then add back to full
-            % variables
-            surge_temp = surge(~isnan(surge));
-            sway_temp = sway(~isnan(sway));
-            heave_temp = heave(~isnan(heave));
-            
-            % Filter all three accel vectors, 5th order butterworth, with
-            % passband between 2 Hz and 15 Hz
-            fny = fs/2;
-            if fs/2 < 15
-                pass = 2;
-                [b,a]=butter(5,pass/fny,'high');
-            else
-                pass = [2, 15];
-                [b,a]=butter(5,pass/fny,'bandpass');
-            end           
-            
-            surge_filt = filtfilt(b,a, surge_temp);
-            sway_filt = filtfilt(b,a, sway_temp);
-            heave_filt = filtfilt(b,a, heave_temp);
-           
-            % Get surge diff
-            surge_diff = diff(surge_filt);
-            sway_diff = diff(sway_filt);
-            heave_diff = diff(heave_filt);
-
-            % Get Shannon entropy
-            surge_se = log(abs(surge_diff))*sum(abs(surge_diff));
-            sway_se = log(abs(sway_diff))*sum(abs(sway_diff));
-            heave_se = log(abs(heave_diff))*sum(abs(heave_diff));
-
-            % Get smoothed Shannon entropy - over 5 second window
-            surge_smooth = movmean(surge_se, 5*fs);
-            sway_smooth = movmean(sway_se, 5*fs);
-            heave_smooth = movmean(heave_se, 5*fs);
-            
-            % Calculate jerk using njerk
-            jerk_filt = njerk([surge_filt, sway_filt, heave_filt], fs);
-            
-            % Get Shannon entropy of jerk
-            jerk_se = log(abs(jerk_filt))*sum(abs(jerk_filt));
- 
-            % Get smoothed Shannon entropy
-            jerk_smooth = movmean(jerk_se', 5*fs); 
-            
-            % Build filter for prh
-            fny = fs/2;
-            
-            % Make temp variables to work with and then add back to full
-            % variables
-            pitch_temp = pitch(~isnan(pitch));
-            roll_temp = roll(~isnan(roll));
-            head_temp = head(~isnan(head));
-            
-            % Calculate filtered prh signals
-            pitch_filt = filtfilt(b, a, pitch_temp);
-            roll_filt = filtfilt(b, a, roll_temp);
-            head_filt = filtfilt(b, a, head_temp);
-
-            % Get prh diff
-            pitch_diff = diff(pitch_filt);
-            roll_diff = diff(roll_filt);
-            head_diff = diff(head_filt);
-
-            % Get Shannon entropy
-            pitch_se = log(abs(pitch_diff))*sum(abs(pitch_diff));
-            roll_se = log(abs(roll_diff))*sum(abs(roll_diff));
-            head_se = log(abs(head_diff))*sum(abs(head_diff));
-
-            % Get smoothed Shannon entropy
-            pitch_smooth = movmean(pitch_se, 5*fs);
[…]
