### 2025-07-11 770a681
Make function intros consistent
