+++ respdetect/main.m
@@ -68,164 +68,12 @@ taglist = load_data(data_path, 'mn');
 make_metadata(taglist, data_path);
 
 %% Step 3: Find dives
-for k = 1:length(taglist);
-    
-    % Load in tag
-    tag = taglist{k};
-    
-    % Load in metadata
-    metadata = load(strcat(data_path, "\metadata\", tag, "md"));
-    clear tag
-    
-    %Set path for prh files
-    settagpath('prh',strcat(data_path,'\prh'));
-            
-    % Load the existing prh file
-    loadprh(metadata.tag);
-    
-    % Calculate time for entire tag deployment
-    [time_sec, time_min, time_hour] =calc_time(metadata.fs, p);
-    
-    % Subset p to tag on and off times
-    start_idx = find(abs(time_sec-metadata.tag_on)==min(abs(time_sec-metadata.tag_on)));
-    end_idx = find(abs(time_sec-metadata.tag_off)==min(abs(time_sec-metadata.tag_off)));
-    
-    if end_idx == length(time_sec)
-        end_idx = end_idx-1;
-    end
-    p_tag = p(start_idx:end_idx);
-    
-    % Calculate time for entire tag deployment
-    [time_sec, time_min, time_hour] =calc_time(metadata.fs, p_tag);
-    
-    % Make a diving file
-    diving_fname = strcat(data_path, "\diving\", metadata.tag, "dives.mat");
-    if isfile(diving_fname) == 1
-        fprintf("A diving table exists for %s - go the next section!\n", metadata.tag)
-    else
-        fprintf("No diving table  exists for %s.\n", metadata.tag)
-        str = input("Do you want to make a diving table now? (y/n)\n",'s');
-        if strcmp(str, "y") == 1
-            
-            % Set dive threshold and find dives
-            prompt = {'Enter a dive threshold in meters:'};
-            dlgtitle = 'Input';
-            dims = [1 35];
-            definput = {'5'};
-            answer = inputdlg(prompt,dlgtitle,dims,definput);
-            dive_thres = str2num(answer{1});
-            clear prompt dlgtitle dims definput answer
-            
-            T = finddives(p_tag,fs, [dive_thres, 1, 0]);
-            
-            if size(T, 1) <= 1
-                display('Only 1 deep dive! Not continuing analysis...')
-                
-            else
-                
-                % Plot dives
-                plot_dives(T, time_sec, p_tag);
-                
-                % Add back start index if needed to time variables
-                if metadata.tag_on ~=0
-                    T(:, [1:2, 4]) = T(:, [1:2, 4]) + metadata.tag_on;
-                end
-                
-                % Calculate time for entire tag deployment
-                [time_sec, time_min, time_hour] =calc_time(metadata.fs, p);
-                               
-                % Extract dive information from T
-                for i = 1:size(T, 1)
-                    
-                    tag{i} = metadata.tag;
-                    depth_thres(i) = dive_thres;
-                    dive_num(i) = i;
-                    dive_start(i) = T(i, 1);
-                    dive_end(i) = T(i, 2);
-                    max_depth(i) = T(i, 3);
-                    time_maxdepth(i) = T(i, 4);
-                    dive_dur(i) = dive_end(i) - dive_start(i);
-                    
-                    % Get dive start and end in indices
-                    start_idx_dive = find(abs(time_sec-dive_start(i))==min(abs(time_sec-dive_start(i))));
-                    end_idx_dive = find(abs(time_sec-dive_end(i))==min(abs(time_sec-dive_end(i))));
 
-                end
-                
-                % Extract surface information from dive information
-                for i = 1:size(T, 1)-1
[…]
