+++ respdetect/setup_dirs.m
@@ -37,28 +37,28 @@ function [recdir, prefix, acousaud_filename] = setup_dirs(tag, tag_ver, dataPath
     %% Use tag name to assign directories and filenames
     sp_year = tag(1:4);
     % Set parent directory
-    recdir = strcat(data_path, sp_year, '\', tag);
+    recdir = strcat(dataPath, sp_year, '\', tag);
     % Set deployment name, probably the same as the parent folder
     deploy_name = tag;
     prefix = tag;
     speciesCode = tag(1:2);
     
     % Set name of acoustic audit file
-    acousaud_filename = strcat(data_path, speciesCode, '\audit\', tag, '_acousticaud.txt');
+    acousaud_filename = strcat(dataPath, speciesCode, '\audit\', tag, '_acousticaud.txt');
     
     if strcmp(tag_ver, 'D3') == 1
         addpath(genpath(strcat(mat_tools_path, '\d3'))); %Add all of your tools to the path
     elseif strcmp(tag_ver, 'D2') == 1
         addpath(genpath(strcat(mat_tools_path, '\dtag2'))); %Add all of your tools to the path
-        settagpath('audio',strcat(data_path, speciesCode),'cal',strcat(data_path, speciesCode,'\cal'));
+        settagpath('audio',strcat(dataPath, speciesCode),'cal',strcat(dataPath, speciesCode,'\cal'));
     end
     
     %% Set other paths
     %Set path for prh files
-    settagpath('prh',strcat(data_path, speciesCode, '\prh'));
+    settagpath('prh',strcat(dataPath, speciesCode, '\prh'));
     
     %Set path for audit files
-    settagpath('audit',strcat(data_path, speciesCode,'\audit'));
+    settagpath('audit',strcat(dataPath, speciesCode,'\audit'));
     
 end
 
