### 2025-07-11 8bce3cb
Move dive_thres out to script
