### 2025-07-11 6a49354
Making breath detections into functions... through 5e
