+++ respdetect/main.m
@@ -69,7 +69,7 @@ make_metadata(taglist, data_path);
 
 %% Step 3: Find dives
 
-% Set the minimum depth of a dive
+% Set the minimum depth to count a dive
 dive_thres = 5;
 
 make_dives(taglist, data_path, dive_thres);
@@ -80,486 +80,7 @@ make_move(taglist, dataPath)
 
 %% Step 5: Detect breaths
 
-for k = 1:length(taglist)
-    
-    %% Step 5a: Import tag data
-    tag = taglist{k};
-
-    % Load in metadata
-    metadata = load(strcat(data_path, "\metadata\", tag, "md"));
-    clear tag
-    
-    %Set path for prh files
-    settagpath('prh',strcat(data_path,'\prh'));  
-            
-    % Load the existing prh file
-    loadprh(metadata.tag);
-    
-    % Load in movement data
-    load(strcat(data_path, "\movement\", metadata.tag, "movement.mat"));
-    
-    % Load in diving data
-    load(strcat(data_path, "\diving\", metadata.tag, "dives.mat"));
-    load(strcat(data_path, "\diving\", metadata.tag, "divetable.mat"));
-    
-    % Calculate time variables for full tag deployment
-    [time_sec, time_min, time_hour] =calc_time(metadata.fs, p);
-       
-    %% Step 5b: Subset deployment to tag on time only
-    
-    start_idx = find(abs(time_sec-metadata.tag_on)==min(abs(time_sec-metadata.tag_on)));
-    end_idx = find(abs(time_sec-metadata.tag_off)==min(abs(time_sec-metadata.tag_off)));
-    
-    if end_idx == length(time_sec)
-        end_idx = end_idx-1;
-    end
-    
-    % If the tag on time is when the tag is near the surface, we are going to
-    % redefine the start idx as the first time the tag goes to a depth of 5 m
-    % the reason for this being that the tag on 
-    % result in a big jerk spike that will interfere with peak detection 
-    % for breaths
-    
-    if p(start_idx)<5
-        start_idx = find(p(start_idx:end_idx)>=5, 1)+start_idx;
-    end
-    
-    % Subset p to only when tag is on
-    p_tag = p(start_idx:end_idx);
-
-    %% Step 5c: Identify surface periods
-    % Smooth depth signal
-    p_smooth = smoothdata(p, 'movmean', fs);
-    p_smooth_tag = smoothdata(p_tag, 'movmean', fs);
-    p_shallow = p_smooth_tag;
-    
-    % Remove any pressure data that is greater than 0.5 m and get indexes
-    % of shallow periods
-    p_shallow(p_smooth_tag>0.5) = NaN;
-    p_shallow_idx = find(~isnan(p_shallow));
-    
-    % Plot smoothed depth for only time when tag is on
-    figure('units','normalized','outerposition',[0 0 1 1]);
-    p1 = plot(time_min(start_idx:end_idx), p_smooth_tag, 'k', 'LineWidth', 1); hold on
-    set(gca, 'YDir', 'reverse');
-    xlabel('Time (min)'); ylabel('Depth (m)');
-    
-    % Find start and end of surface periods
-    p_shallow_breaks_end = find(diff(p_shallow_idx)>1);
-    p_shallow_breaks_start = find(diff(p_shallow_idx)>1)+1;
-    
-    % Define variable to store surfacings
-    p_shallow_ints = [[1; p_shallow_breaks_start], [p_shallow_breaks_end; length(p_shallow_idx)]];
-    
-    % Make third column which is duration of surfacing in indices
[…]
