### 2025-07-14 9c8c8cc
Added function to allow user to save figure and overwrite or append suffix
