### 2025-07-14 ffed96d
Update set_paths.m
