+++ respdetect/detect_breaths.m
@@ -1,7 +1,10 @@
-function detect_breaths(taglist, dataPath)
+function detect_breaths(taglist, dataPath, n_sec, min_sec_apart, win_sec)
     arguments
         taglist (1, :) cell
         dataPath (1,:) char
+        n_sec (1, 1) double
+        min_sec_apart (1, 1) double
+        win_sec (1, 1) double
     end
     % Uses previously calculated information to identify breaths
     %
@@ -25,22 +28,17 @@ function detect_breaths(taglist, dataPath)
         %% Step 5a: Import tag data
         tag = taglist{k}; speciesCode = tag(1:2);
         
-        % Load in metadata
-        metadata = load(strcat(dataPath, speciesCode, "\metadata\", tag, "md"));
-        clear tag
-        
+        % Load in metadata, movement, and diving data
+        load(strcat(dataPath, speciesCode, "\metadata\", tag, "md")); clear tag;
+        load(strcat(dataPath, speciesCode, "\movement\", metadata.tag, "movement.mat"));
+        load(strcat(dataPath, speciesCode, "\diving\", metadata.tag, "dives.mat"));
+        load(strcat(dataPath, speciesCode, "\diving\", metadata.tag, "divetable.mat"));
+              
         %Set path for prh files
         settagpath('prh',strcat(dataPath, speciesCode, '\prh'));
         
         % Load the existing prh file
         loadprh(metadata.tag);
-        
-        % Load in movement data
-        load(strcat(dataPath, speciesCode, "\movement\", metadata.tag, "movement.mat"));
-        
-        % Load in diving data
-        load(strcat(dataPath, speciesCode, "\diving\", metadata.tag, "dives.mat"));
-        load(strcat(dataPath, speciesCode, "\diving\", metadata.tag, "divetable.mat"));
         
         % Calculate time variables for full tag deployment
         [time_sec, time_min, time_hour] =calc_time(metadata.fs, p);
@@ -49,18 +47,17 @@ function detect_breaths(taglist, dataPath)
         
         [p_tag, start_idx, end_idx] = subset_p(metadata, p, time_sec);
         
-        %% Step 5c: Identify surface periods
+        %% Step 5c: Identify surface periods & classify them
         
-        [p_shallow_ints, p_shallow_idx, p_shallow] = get_shallowints(metadata, p, p_tag);
+        [p_shallow_ints, p_shallow_idx, p_shallow, p_smooth_tag] = get_shallowints(metadata, p, p_tag);
         
         % If these periods are less than 10 seconds then we say they are a "single
         % breath surfacing" otherwise they are a "logging surfacings"
-        n_sec = 10;
-        [single_breath_surf_rows, logging_surf_rows, logging_start_idxs, logging_end_idxs, logging_ints_s, p_shallow_ints] = ...
+        [single_breath_surf_rows, logging_surf_rows, logging_start_idxs, logging_end_idxs, logging_ints_s, p_shallow_ints, all_breath_locs] = ...
             classify_surfs(p_shallow_ints, p_shallow_idx, p_shallow, time_sec, start_idx, n_sec, metadata.fs);
         
         %% Step 5d: Start plot of surfacing periods
-       
+        
         [p1, p2, p3, p4, p5] = plot_surfs(time_min, p_tag, p_smooth_tag, start_idx, end_idx, ...
             p_shallow_idx, p_shallow_ints, p_shallow, ...
             logging_surf_rows, single_breath_surf_rows, metadata);
@@ -74,74 +71,20 @@ function detect_breaths(taglist, dataPath)
[…]
