### 2025-07-14 3cc5086
Write functions for breath detection
