+++ respdetect/get_logbreaths.m
@@ -1,0 +1,133 @@
+function [all_breath_locs] = get_logbreaths( ...
+        val3, temp_diff_break, j_wins, s_wins, p_wins, ...
+        j_wins_breaks, s_wins_breaks, p_wins_breaks, ...
+        p_smooth_tag, fs, start_idx, all_breath_locs)
+    arguments
+        val3 (:, 1) double
+        temp_diff_break (:, 1) double
+        j_wins (1, :) double
+        s_wins (1, :) double
+        p_wins (1, :) double
+        j_wins_breaks (1, :) double
+        s_wins_breaks (1, :) double
+        p_wins_breaks (1, :) double
+        p_smooth_tag (1, :) double
+        fs (1, 1) double {mustBePositive}
+        start_idx (1, 1) double {mustBeInteger, mustBeNonnegative}
+        all_breath_locs struct
+    end
+    % Detects breath locations during logging surfacings.
+    %
+    % Inputs:
+    %   val3              - Vector of indices where at least 2 of 3 peak signals overlap
+    %   temp_diff_break   - Indices where continuous regions in val3 are interrupted
+    %   j_wins            - Jerk event window indices
+    %   s_wins            - Surge event window indices
+    %   p_wins            - Pitch event window indices
+    %   j_wins_breaks     - End indices of each jerk window (used to detect duplicates)
+    %   s_wins_breaks     - End indices of each surge window
+    %   p_wins_breaks     - End indices of each pitch window
+    %   p_smooth_tag      - Smoothed pressure signal used to filter valid breathing events
+    %   fs                - Sampling frequency (Hz)
+    %   start_idx         - Offset from beginning of full pressure signal
+    %   all_breath_locs   - Struct with previous breath detections (.breath_idx and .type)
+    %
+    % Output:
+    %   all_breath_locs   - Updated struct with new logging breath detections added and cleaned
+    %
+    % Usage:
+    %   all_breath_locs = identify_logging_breaths(val3, temp_diff_break, ...
+    %       j_wins, s_wins, p_wins, ...
+    %       j_wins_breaks, s_wins_breaks, p_wins_breaks, ...
+    %       p_smooth_tag, fs, start_idx, all_breath_locs)
+    %
+    % Author: Ashley Blawas
+    % Last Updated: 7/11/2025
+    % Stanford University
+    
+    log_breath_locs = [];
+    if ~isempty(temp_diff_break)
+        for c = 1:(length(temp_diff_break)+1)
+            % Initialize for first loop
+            if c == 1
+                j_win_count = 0; s_win_count = 0; p_win_count = 0;
+                cont_val3_prev = -3 * fs; % Set pre-window for breath separation
+                cont_range = 1:temp_diff_break(1);
+            elseif c == length(temp_diff_break)+1
+                cont_val3_prev = cont_val3;
+                cont_range = temp_diff_break(end)+1:length(val3);
+            else
+                cont_val3_prev = cont_val3;
+                cont_range = temp_diff_break(c-1)+1:temp_diff_break(c);
+            end
+            
+            if length(cont_range) > 1 * fs
+                cont_val3 = val3(cont_range);
+                
+                % Track window overlap with jerk, surge, and pitch
+                cond = 0;
+                
+                % Jerk
+                j_temp_int = intersect(cont_val3, j_wins);
+                j_win_count_prev = j_win_count;
+                if ~isempty(j_temp_int)
+                    j_win_count = find(j_temp_int(end) <= j_wins_breaks, 1);
+                    cond = 1;
+                end
+                
+                % Surge
+                s_temp_int = intersect(cont_val3, s_wins);
[…]

[…]
