### 2025-07-14 647b6c5
Make function plot_breaths
