+++ respdetect/main.m
@@ -93,106 +93,9 @@ detect_breaths(taglist, data_path, n_sec, min_sec_apart, win_sec)
 %% Step 6: Plot all breaths
 
 for k = 1:length(taglist)
-    %% Step 6a: Load in tag data
-    tag = taglist{k};
     
-    %Load in metadata
-    metadata = load(strcat(data_path, "\metadata\", tag, "md"));
-    clear tag
-    
-    %Set path for prh files
-    settagpath('prh',strcat(data_path,'\prh'));
-    
-    % Load the existing prh file
-    loadprh(metadata.tag);
-       
-    % Load in movement data
-    load(strcat(data_path, "\movement\", metadata.tag, "movement.mat"), 'jerk_smooth', 'surge_smooth', 'pitch_smooth');
-    
-    % Load in breathing information
-    load(strcat(data_path, "\breaths\", metadata.tag, "breaths.mat"));
-    
-    [time_sec, time_min, time_hour] =calc_time(metadata.fs, pitch); %Recalculate time
-    
-    % Load in breaths
-    breath_idx = all_breath_locs.breath_idx;
-    
-    if strcmp(metadata.tag_ver, "CATS") == 1
-        breath_times =  datetime(DN(all_breath_locs.breath_idx), 'ConvertFrom', 'datenum', 'Format', 'yyyy-MM-dd HH:mm:ss.SSS');
+  plot_breaths(data_path, taglist, k);
         
-        % Create datetime variable
-        date = datetime(DN, 'ConvertFrom', 'datenum', 'Format', 'yyyy-MM-dd HH:mm:ss.SSS');
-        xval = date;
-    else
-        breath_times = time_min(all_breath_locs.breath_idx);
-        xval = time_min;
-    end
-    
-   [breath_times, sortidx]  = sort(breath_times);
-   breath_type = all_breath_locs.type(sortidx, :); 
-    
-    %% Step 6b: Plot all breaths
-    
-    % Filter signals to tag on time for plotting
-    start_idx = find(abs(time_sec-metadata.tag_on)==min(abs(time_sec-metadata.tag_on)));
-    end_idx = find(abs(time_sec-metadata.tag_off)==min(abs(time_sec-metadata.tag_off)));
-    
-    p_smooth(1:start_idx) = NaN;
-    p_smooth(end_idx:length(p_smooth)) = NaN;
-    
-    jerk_smooth(1:start_idx) = NaN;
-    jerk_smooth(end_idx:length(jerk_smooth)) = NaN;
-    
-    surge_smooth(1:start_idx) = NaN;
-    surge_smooth(end_idx:length(surge_smooth)) = NaN;
-    
-    pitch_smooth(1:start_idx) = NaN;
-    pitch_smooth(end_idx:length(pitch_smooth)) = NaN;
-    
-    figure
-    title(metadata.tag, 'Interpreter', 'none');
-    ax(1)=subplot(4, 1, 1);
-    plot(xval, p_smooth, 'k', 'LineWidth', 1.5); hold on
-    set(gca,'Ydir','reverse')
-    ylabel('Depth (m)');
-    
-    hold on
-    scatter(breath_times(breath_type == 'ss'), p_smooth(breath_idx(breath_type == 'ss')), 60, 'cs', 'filled', 'MarkerEdgeColor', 'k', 'MarkerFaceAlpha', .75, 'MarkerEdgeAlpha', .75)
[…]
