### 2025-07-14 ashleyblawas/respdetect#1 @ashleyblawas
Thank you for these suggestions - I have restructured this main.m file into many separate functions and now have them ordered in a file called run_all.m where the user can define the paths and variables for a given experiment.
