### 2025-07-14 ashleyblawas/respdetect#3 @ashleyblawas
I have gone through and added input validation to all functions using Matlab's argument validation system - thank you for bringing this to my attention!
