### 2025-07-14 ashleyblawas/respdetect#4 @ashleyblawas
Thanks for the reminder on this. I have gone through and documented each function with a description, inputs, outputs, and usage. 👍
