### 2025-07-14 ashleyblawas/respdetect#13 @ashleyblawas
Thanks for catching this - to address this issues I have chosen to rename the `nanmedian` function in the `dtagtools` packaged within `respdetect` to `nanmedian2` which should eliminate the problem.
