### 2025-07-14 ashleyblawas/respdetect#5 @ashleyblawas
In re-writing the script into functions I have tried to address all of these instances... a single digit number remain related to importing specific variables from files as opposed to the whole file but these are infrequent and do not generate conflicts.
