+++ respdetect/save_fig.m
@@ -23,8 +23,8 @@ function save_fig(dataPath, speciesCode, metadata, fig_suffix)
                   'Options:\n' ...
                   '  [o] Overwrite existing file\n' ...
                   '  [a] Append custom suffix\n' ...
-                  '  [n] Cancel save\n' ...
-                  'Choose an option (o/a/n): ']);
+                  '  [c] Cancel save\n' ...
+                  'Choose an option (o/a/c): ']);
               txt = input(prompt, "s");
         
         switch lower(txt)
