### 2025-07-14 ashleyblawas/respdetect#14 @ashleyblawas
This is a really helpful functionality, thanks for this idea. I have gone through and added this functionality for the metadata, dive, and movement file making
