### 2025-07-14 ashleyblawas/respdetect#8 @ashleyblawas
Apologizes... this was describing some versions of this code which had additional functions to export the data to R and do other validation steps with supplementary datasets (i.e., acoustics). Those have not been fully deployed here and thus I have removed them from the Wiki. Sorry for the confusion!
