### 2025-07-14 ashleyblawas/respdetect#8 @ashleyblawas
To clarify - because some of these functionalities are in development I instead decided to collapse them into a Wiki page called In development so someone could reach out if any of these were relevant to their work... otherwise I have deleted these additional pages.
