### 2025-07-14 ashleyblawas/respdetect#15 @ashleyblawas
I have addressed many of these and will get to the rest tomorrow - thank you!
