### 2025-07-14 146c8bf
Resolve conflicts when reverted load_data deletion
