### 2025-07-14 f3ee49a
Update plot_dives.m
