+++ respdetect/run_all.m
@@ -102,6 +102,9 @@ end
 
 %% Step 7a: Test detections for gm08_143b
 % For these detections I used the following parameters:
+%   Tag on: 201.40 seconds (default)
+%   Tag off: 15835.21 seconds (manual)
+%   Tag version: D2 (default)
 %
 % Set the minimum depth to count a dive
 %   dive_thres = 5;
@@ -118,14 +121,14 @@ end
 clear; clc; close all;
 
 % Load in validation breathing information
-load(strcat(tools_path, "\tests\gm\gm08_143bbreaths_val.mat"));
+load(strcat(data_path, "\gm\gm08_143bbreaths_val.mat"));
 
 % Load in breaths
 true_breath_idx = all_breath_locs.breath_idx;
 true_breath_s = true_breath_idx/fs;
 
 % Load in breathing information generated by user
-load(strcat(tools_path, "\tests\gm\breaths\gm08_143bbreaths.mat"));
+load(strcat(data_path, "\gm\breaths\gm08_143bbreaths.mat"));
 
 % Load in breaths
 breath_idx = all_breath_locs.breath_idx;
@@ -139,17 +142,32 @@ fprintf('The mean difference between your detections and the true detections is:
 fprintf('The max difference between your detections and the true detections is: %0.1f seconds\n', max(breath_idx_diff)); 
 
 %% Step 7b: Test detections for mn17_310a
-clearvars -except taglist tools_path mat_tools_path data_path; clc; close all
+% For these detections I used the following parameters:
+%   Tag on: 201.40 seconds (default)
+%   Tag off: 15835.21 seconds (manual)
+%   Tag version: D2 (default)
+%
+% Set the minimum depth to count a dive
+%   dive_thres = 5;
+%
+% The # of seconds to distinguish between single-breath surfacing and logging surfacing
+%   n_sec = 10;
+%
+% The minimum # of seconds between peaks detected in movement signals during logging period.
+%   min_sec_apart = 3;
+%
+% The window size (# of sec) for finding aligning peaks in jerk, surge, and pitch.
+%   win_sec = 5;
 
 % Load in validation breathing information
-load(strcat(tools_path, "\tests\mn\mn17_310abreaths_val.mat"));
+load(strcat(data_path, "\tests\mn\mn17_310abreaths_val.mat"));
 
 % Load in breaths
 true_breath_idx = all_breath_locs.breath_idx;
 true_breath_s = true_breath_idx/fs;
 
 % Load in breathing information generated by user
-load(strcat(tools_path, "\tests\breaths\mn17_310abreaths.mat"));
+load(strcat(data_path, "\tests\breaths\mn17_310abreaths.mat"));
 
 % Load in breaths
 breath_idx = all_breath_locs.breath_idx;
