### 2025-07-14 2b42e87
Fix detect_breaths to work with no logging breaths
