+++ respdetect/run_all.m
@@ -60,7 +60,7 @@ set_paths(data_path);
 
 % Allow the user to select the prh files to analyze from a given species
 % folder
-taglist = load_data(data_path, 'gm');
+taglist = load_data(data_path, 'mn');
 
 %% Step 2: Make metadata file
 
@@ -121,7 +121,7 @@ end
 clear; clc; close all;
 
 % Load in validation breathing information
-load(strcat(data_path, "\gm\gm08_143bbreaths_val.mat"));
+load(strcat(data_path, "\gm\breaths\gm08_143bbreaths_val.mat"));
 
 % Load in breaths
 true_breath_idx = all_breath_locs.breath_idx;
@@ -143,8 +143,8 @@ fprintf('The max difference between your detections and the true detections is: 
 
 %% Step 7b: Test detections for mn17_310a
 % For these detections I used the following parameters:
-%   Tag on: 201.40 seconds (default)
-%   Tag off: 15835.21 seconds (manual)
+%   Tag on: 0.0 seconds, i.e., on at start (default)
+%   Tag off:  24322.00 seconds (default)
 %   Tag version: D2 (default)
 %
 % Set the minimum depth to count a dive
@@ -160,14 +160,14 @@ fprintf('The max difference between your detections and the true detections is: 
 %   win_sec = 5;
 
 % Load in validation breathing information
-load(strcat(data_path, "\tests\mn\mn17_310abreaths_val.mat"));
+load(strcat(data_path, "\mn\breaths\mn17_310abreaths_val.mat"));
 
 % Load in breaths
 true_breath_idx = all_breath_locs.breath_idx;
 true_breath_s = true_breath_idx/fs;
 
 % Load in breathing information generated by user
-load(strcat(data_path, "\tests\breaths\mn17_310abreaths.mat"));
+load(strcat(data_path, "\mn\breaths\mn17_310abreaths.mat"));
 
 % Load in breaths
 breath_idx = all_breath_locs.breath_idx;
