### 2025-07-14 ced048d
Built testing into functions & improved information provided from test
