+++ respdetect/run_all.m
@@ -60,7 +60,7 @@ set_paths(data_path);
 
 % Allow the user to select the prh files to analyze from a given species
 % folder
-taglist = load_data(data_path, 'mn');
+taglist = load_data(data_path, 'gm');
 
 %% Step 2: Make metadata file
 
@@ -118,28 +118,10 @@ end
 % The window size (# of sec) for finding aligning peaks in jerk, surge, and pitch.
 %   win_sec = 5;
 
-clear; clc; close all;
+% Tolerance for matching detections (in seconds)
+tolerance = 0.2;
 
-% Load in validation breathing information
-load(strcat(data_path, "\gm\breaths\gm08_143bbreaths_val.mat"));
-
-% Load in breaths
-true_breath_idx = all_breath_locs.breath_idx;
-true_breath_s = true_breath_idx/fs;
-
-% Load in breathing information generated by user
-load(strcat(data_path, "\gm\breaths\gm08_143bbreaths.mat"));
-
-% Load in breaths
-breath_idx = all_breath_locs.breath_idx;
-breath_s = breath_idx/fs;
-
-% Calculate similarity between true and users detections
-breath_idx_diff = true_breath_s - breath_s;
-
-% Print out similarity metrics
-fprintf('The mean difference between your detections and the true detections is: %0.1f seconds\n', mean(breath_idx_diff)); 
-fprintf('The max difference between your detections and the true detections is: %0.1f seconds\n', max(breath_idx_diff)); 
+test_breaths(data_path, "gm08_143bbreaths_val.mat", "gm08_143bbreaths.mat", tolerance)
 
 %% Step 7b: Test detections for mn17_310a
 % For these detections I used the following parameters:
@@ -159,23 +141,7 @@ fprintf('The max difference between your detections and the true detections is: 
 % The window size (# of sec) for finding aligning peaks in jerk, surge, and pitch.
 %   win_sec = 5;
 
-% Load in validation breathing information
-load(strcat(data_path, "\mn\breaths\mn17_310abreaths_val.mat"));
+% Tolerance for matching detections (in seconds)
+tolerance = 0.2;
 
-% Load in breaths
-true_breath_idx = all_breath_locs.breath_idx;
-true_breath_s = true_breath_idx/fs;
-
-% Load in breathing information generated by user
-load(strcat(data_path, "\mn\breaths\mn17_310abreaths.mat"));
-
-% Load in breaths
-breath_idx = all_breath_locs.breath_idx;
-breath_s = breath_idx/fs;
-
-% Calculate similarity between true and users detections
-breath_idx_diff = true_breath_s - breath_s;
-
-% Print out similarity metrics
-fprintf('The mean difference between your detections and the true detections is: %0.1f seconds\n', mean(breath_idx_diff)); 
-fprintf('The max difference between your detections and the true detections is: %0.1f seconds\n', max(breath_idx_diff)); 
+test_breaths(data_path, "mn17_310abreaths_val.mat", "mn17_310abreaths.mat", tolerance)
