### 2025-07-14 ashleyblawas/respdetect#15 @ashleyblawas
I have address all of these comments including improving the testing of this package to make it more clear where the user may have gone wrong. I have also added many fail safes in the functions to try and prevent issues prior to testing.

The challenge with testing is that outcome of the breath detection code is fairly binary—either breaths will be detected at all the indicies that correspond to the depth/kinematic peaks or some error will preclude a set of detections (whether single surfacing breath or logging breaths) entirely. This makes it challenging to write automated test code or to pinpoint the exact source of error at the testing/validation stage.

Due to this, the testing here focuses more on comparing summary statistics and matching breaths within a given tolerance to assess whether differences are meaningful or just artifacts of alignment.
