### 2025-07-14 ashleyblawas/respdetect#6 @ashleyblawas
I have updated the testing to be more robust and informative. I have also addressed this issue in #15 in more detail. If you have any other suggestions for how to improve this testing I am open to these suggestions.
