### 2025-07-22 #7858 @kthyng
@sfregosi and @smnnlt: @ashleyblawas has commented on the issues you raised for this review. Can you follow up on them and see what you think?

@ashleyblawas In this situation, the reviewers should decide whether or not to close the issues they opened to reflect if they think it is finished or not. It is possible that not every issue will end up closed, depending on how important it is to do as part of this review. Please reopen the issues so we can track what are the outstanding issues, thanks.
