### 2025-08-07 ashleyblawas/respdetect#4 @smnnlt
I appreciate your work on this, but I think there are still some cases where the documentation does not the match the function. E.g., https://github.com/ashleyblawas/respdetect/blob/f300d79058af914fd6f60b09acb205226a7c4f30/respdetect/calc_move.m#L1-L17
Please check every function again.
