### 2025-08-07 ashleyblawas/respdetect#2 @smnnlt
I like the implemetation of path selection as part of the function. Is it possible to make it possible to also use relative paths as an input for the `set_paths()` function? This would also allow users to run the file with the example data without needing to modify the `data_path` variable in `run_all.m`
