### 2025-08-07 ashleyblawas/respdetect#5 @smnnlt
Good job, most warnings were eliminated. But some warnings remain, e.g., `detect_peaks.m` shows 7 warnings (all unrelated to importing from files). I strongly suggest to fix all remaining warnings in this and the other files.
