### 2025-08-07 ashleyblawas/respdetect#9 @smnnlt
Thanks! Please just mention the new contribution guidelines in the `README.md`.
