### 2025-08-07 ashleyblawas/respdetect#1 @smnnlt
Good job, I think the script is now easier to use and modify for users.
