### 2025-08-07 ashleyblawas/respdetect#16 @smnnlt
Bug in make_metadata()
I think there is a bug in make_metadata():
https://github.com/ashleyblawas/respdetect/blob/f300d79058af914fd6f60b09acb205226a7c4f30/respdetect/make_metadata.m#L72
Which should be `settagpath('prh',strcat(dataPath, '\', speciesCode,'\prh')); `

Comment for the JOSS review (https://github.com/openjournals/joss-reviews/issues/7858).
