### 2025-08-07 ashleyblawas/respdetect#16 @smnnlt
Same here: https://github.com/ashleyblawas/respdetect/blob/f300d79058af914fd6f60b09acb205226a7c4f30/respdetect/make_dives.m#L40
