### 2025-08-11 ashleyblawas/respdetect#2 @ashleyblawas
Thanks for the suggestion for the set_paths() function. This now allows the user to leave the path blank and it then defaults to the path where the test file exist.
