### 2025-08-11 1ab858d
Update set_paths to have a default if left blank
