+++ respdetect/set_paths.m
@@ -1,16 +1,18 @@
 function set_paths(dataPath)
     arguments
-        dataPath (1,:) char
+        dataPath (1,:) char = ''  % Make input optional
     end
     % Adds respdetect, dtagtools, and user-defined data path by taking in
     % the data path and the tool path.
     %
     % Inputs:
     %   dataPath - Base path to data (e.g., 'C:\Documents\my_data\')
-    %   toolPath - Base path to data (e.g., 'C:\Documents\toolboxes\respdetect')
+    %              If empty or not provided, defaults to ../tests from here
     %
     % Usage:
-    %   set_paths('C:/Users/YourName/Documents/data')
+    %   set_paths();               % Uses default tests directory
+    %   set_paths('');            % Same as above
+    %   set_paths('path/to/data') % Uses user-specified path
     %
     % Assumptions:
     %   - This function is located somewhere inside the respdetect directory.
@@ -25,29 +27,41 @@ function set_paths(dataPath)
     currentFile = mfilename('fullpath');
     [currentDir, ~, ~] = fileparts(currentFile);
     
-    % Find 'respdetect' root
-    respdetectPath = locate_folder_up(currentDir, 'respdetect');
-    if isempty(respdetectPath)
-        error('Could not locate the "respdetect" root folder.');
+    % Go one level up to reach repo root
+    repoRoot = fileparts(currentDir);             % .../ (repo root)
+    
+    % Define paths
+    respdetectPath = fullfile(repoRoot, 'respdetect');
+    dtagtoolsPath  = fullfile(respdetectPath, 'dtagtools');
+
+    % Resolve data path
+    if isempty(dataPath)
+        dataPath = fullfile(repoRoot, 'tests');
+    elseif ~isfolder(dataPath)
+        % Assume it's a relative path, resolve it from current working dir
+        dataPath = fullfile(pwd, dataPath);
     end
     
     % Define dtagtools path relative to respdetect
     dtagtoolsPath = fullfile(respdetectPath, 'dtagtools');
     
-    % Check if folders exist
+    % Validate paths
+    if ~isfolder(respdetectPath)
+        error('respdetect path not found at: %s', respdetectPath);
+    end
     if ~isfolder(dtagtoolsPath)
-        error('The dtagtools path does not exist at: %s. Please move dtagtools here.', dtagtoolsPath);
+        error('dtagtools path not found at: %s', dtagtoolsPath);
     end
     if ~isfolder(dataPath)
-        error('The data path does not exist here: %s', dataPath);
+        error('The data path does not exist: %s', dataPath);
     end
-    
+
     % Add paths to MATLAB
     addpath(genpath(respdetectPath));
     addpath(genpath(dtagtoolsPath));
     addpath(dataPath);  % Flat add for data
-    
-    % Display status
+
+    % Display confirmation
     fprintf('Added respdetect path: %s\n', respdetectPath);
     fprintf('Added dtagtools path:  %s\n', dtagtoolsPath);
     fprintf('Added data path:       %s\n', dataPath);
