### 2025-08-11 ashleyblawas/respdetect#2 @ashleyblawas
The `README.md` has also been updated.
