### 2025-08-11 884817c
Updating fxn docs
