### 2025-08-11 797a610
Updating fxn docs and adding argument validation
