### 2025-08-11 9d10c2e
Update fxn documentation and argument val
