### 2025-08-11 82c774c
Updating fxn documentation
