+++ respdetect/plot_surfs.m
@@ -2,44 +2,64 @@ function [p1, p2, p3, p4, p5] = plot_surfs(time_min, p_smooth_tag, start_idx, en
         p_shallow_idx, p_shallow_ints, p_shallow, ...
         logging_surf_rows, single_breath_surf_rows, metadata)
     arguments
-        time_min (1, :) double
-        p_smooth_tag (:, 1) double
-        start_idx (1, 1) double
-        end_idx (1, 1) double
-        p_shallow_idx (:, 1) double
-        p_shallow_ints (:, 4) double
-        p_shallow (:, 1) double
-        logging_surf_rows (:, 1) double
-        single_breath_surf_rows (:, 1) double
-        metadata (1, 1) struct
+        time_min (1, :) double                             % Time vector in minutes (row vector)
+        p_smooth_tag (:, 1) double                         % Smoothed depth data
+        start_idx (1, 1) double {mustBeInteger, mustBePositive}  % Start index for plotting
+        end_idx (1, 1) double {mustBeInteger, mustBePositive}    % End index for plotting
+        p_shallow_idx (:, 1) double {mustBeInteger}        % Indices of shallow depth points
+        p_shallow_ints (:, 4) double                       % Matrix of shallow interval start/end indices and metrics
+        p_shallow (:, 1) double                            % Vector with shallow depths (NaNs elsewhere)
+        logging_surf_rows (:, 1) double {mustBeInteger}    % Row indices in `p_shallow_ints` marking logging surfacings
+        single_breath_surf_rows (:, 1) double {mustBeInteger} % Row indices in `p_shallow_ints` marking single-breath surfacings
+        metadata (1, 1) struct                             % Metadata struct containing tag info
     end
-    % Plots surfacings of the two different types
+    
+    % PLOTS_SURFS Visualizes shallow surfacing intervals on a smoothed depth trace.
+    %
+    % This function generates a figure showing smoothed depth data and overlays
+    % two types of surfacing events (logging and single-breath) along with their
+    % start and end points.
     %
     % Inputs:
+    %   time_min                - Time vector in minutes
+    %   p_smooth_tag            - Smoothed depth signal
+    %   start_idx               - Start index for plotting window
+    %   end_idx                 - End index for plotting window
+    %   p_shallow_idx           - Indices where the animal was considered shallow
+    %   p_shallow_ints          - Nx4 matrix of shallow interval information
+    %   p_shallow               - Depth signal with NaNs outside shallow intervals
+    %   logging_surf_rows       - Indices of logging surfacing rows in `p_shallow_ints`
+    %   single_breath_surf_rows - Indices of single-breath surfacing rows in `p_shallow_ints`
+    %   metadata                - Struct with tag metadata (e.g., metadata.tag for title)
     %
     % Outputs:
-    %   p1 - main depth plot
-    %   p2 - logging surfacings
-    %   p3 - single-breath surfacings
-    %   p4 - surfacing start markers
-    %   p5 - surfacing end markers
+    %   p1 - Line object for smoothed depth plot
+    %   p2 - Line object for logging surfacings (magenta)
+    %   p3 - Line object for single-breath surfacings (cyan)
+    %   p4 - Marker object for surfacing start points (green asterisks)
+    %   p5 - Marker object for surfacing end points (red asterisks)
     %
     % Usage:
[…]
