### 2025-08-11 b6af284
Updating fxn doc
