### 2025-08-11 ashleyblawas/respdetect#4 @ashleyblawas
I have double checked and updated all functions with correct function documentation.
