+++ respdetect/calc_move.m
@@ -99,9 +99,6 @@ function calc_move(fs, Aw, p, pitch, roll, head, movement_fname)
     % Get smoothed Shannon entropy
     jerk_smooth = movmean(jerk_se', 5*fs);
     
-    % Build filter for prh
-    fny = fs/2;
-    
     % Make temp variables to work with and then add back to full
     % variables
     pitch_temp = pitch(~isnan(pitch));
@@ -172,6 +169,6 @@ function calc_move(fs, Aw, p, pitch, roll, head, movement_fname)
         'pitch_se', 'head_se', 'roll_se',...
         'pitch_smooth', 'head_smooth', 'roll_smooth');
     
-    display('Movement information calculation complete!');
+    disp('Movement information calculation complete!');
     
 end
