### 2025-08-11 ashleyblawas/respdetect#5 @ashleyblawas
Thanks for catching these. I have gone ahead and cleared up all warnings except for one type (<6 across all files)related to loading individual variables from mat file to avoid conflicts with variables or functions on the path. Because I am generally loading +15 vars from these mat files I have kept these lines to allow all variables to be loaded into the workspace.
