### 2025-08-11 ashleyblawas/respdetect#9 @ashleyblawas
The `README.md` has been updated with a link to `CONTRIBUTING.md`.
