### 2025-08-20 ashleyblawas/respdetect#16 @smnnlt
There may still be issues here:
https://github.com/ashleyblawas/respdetect/blob/202d577ecf202bc199e1c9568f726d74d21b4605/respdetect/make_metadata.m#L69

And here:
https://github.com/ashleyblawas/respdetect/blob/202d577ecf202bc199e1c9568f726d74d21b4605/respdetect/make_metadata.m#L89
