### 2025-08-20 ashleyblawas/respdetect#2 @smnnlt
Thanks for the modification. If I leave `data_path = []` this gives me an error at the moment.
