### 2025-08-25 ashleyblawas/respdetect#16 @ashleyblawas
Hmmm, somehow I didn't run to any errors running it this way on my machine but I see what you mean and have added in the addition of a `"\"` to the path before `speciesCode`.
